Update class for vectors
The class for vectors has been updated in a non backward compatible
way.  In order for it to emulate the SymPy IndexedBase class better,
the differentiation of base and label is introduced.  Basically the
class attempts to unify IndexedBase and Indexed via currying.

Also, internally, direct access of member fields is preferred over the
public properties.  This could give some performance improvements.