Add free variable cache to tensors
This also brings some slight insignificant change in the semantics of
the code.  Now the free variables in expressions are almost always
assumed to be invariant during the course of computation.  In this way,
significant performance boost is possible by skipping all the gathering
of free variables.  When some of them are indeed cancelled, the dummy
reset method can be explicitly called.