Rework the delta simplification core
The delta simplification in terms has been significantly reworked.  On
the users side, the treatment of Dirac delta is removed.  It is not used
much but adds significant performance penalty and code complexity.
Internally, first it is significantly optimized by using flat dictionary
rather than serial sequence for the substitutions.  And the resolution
and composition of simplified deltas are made public to be usable in
other parts of the code.