Update delta simplification result composition
Now the new substitution to be composed with existing ones will be an
iterable of possibly multiple substitutions.  Although one delta in the
expression tree can at most given one substitution, this could give the
function the flexibility to be used in different more complex
situations.

Also the documentation to this function is augmented.