Add delta handling during Wick expansion
Previously the Wick expansion part and the delta simplification are two
orthogonal parts without interaction.  Since explicit consideration of
deltas during the Wick expansion process could possibly save a lot of
computation, now the handling of deltas is introduced to the Wick
expansion code.

The reason for this change is mostly twofold,

1. The performance gain could make it worth doing.

2. By previous refactorization of the delta handling, the code
duplication is made minimal and high-level.