Implement new site-dependent lattice commutation rules
The previous definition has serious flaw that it does not play well with
the serialization method of broadcast variable in Spark.  In this new
design, rather than allowing generic callable objects, the
site-dependent commutation rules can be input by giving a serializable
mathematical expression.

This also has the additional benefit of actually making the interface
more intuitive for non-programmers.