Enhance amplitude summation simplification
Previously, when a summation is not involved in the vectors, it can only
be simplified when it is not actually involved in the amplitude as well.
Now this style of simplification is much expanded.  A tunable call-back
can be used to attempt to simplify a summation over the factors in the
amplitude with it, when it is not involved in the vector part.  When a
simplification is found, it will be used to replace the summation of the
involving factors.  It can be very useful for a lot of applications.