Update range hash and equality comparison
Now we use the label only for hash and equality comparison.  This is to
facilitate the usage of other dummy variables in the ranges of some
summations.

This is in pricinple a backward incompatible change.  However, since it
has been warned that ranges with the same label but different bounds
should not be used, this does not actually break much previous code.