Fix merging with sums over variable bounds
Previously, the merge of terms is decided from the default equality
comparison of the ranges.  Now with ranges being compared solely based
on their label, this behaviour could given errorneous results.  Now it
is fixed.

Also a regression test is added for a possible case where it can go
wrong.