Update default handling of bounds of ranges
In the mapping of scalar functions to terms, the bounds of the
summations are, by default, skipped for performance reasons.  Now it
appears that its judicious turning-on can be extremely difficult on the
caller's side, it is now turned on by default.

This can give a performance penalty on most use cases where we would
never touch the bounds of the ranges.  Hopefully it will not be a big
problem.