Rework internal summation simplification
The simplification of the internal summations are almost reworked to be
able to handle more complex cases.  Now instead of aggregating
summations involved by the same set of factors, all subsets of possibly
summable summations are attempted.  This is accelerated by requiring
rules to give the number of summation indices handled by them
explicitly.

Since now the rules are no longer applied to disjoint sets of factors,
we now have a main loop that attempts the simplifications until none of
the rules is applicable.  The organization is attempted to make the code
run fast.