Revert commit d1bfd91 to fix handling of dummies in bounds
Commit d1bfd91 was made from a good intent to consider the renaming of
dummies appearing in summation bounds.  However, this task will be done
again later when the dummy-renaming substitution is mapped back to all
the scalars in the term.  Doing it twice gives erroneous results that
does not make sense.  In this commit, the task is decided to be
performed later in the treatment of all the scalars.  This is also
consistent with the convention of including the summation bounds in the
scalars of the term and excluding the atomic dummies.