Improve treatment of relations among summations
Previously, in the canonicalization procedure, the relations among the
summations were plainly ignored.  No edge was ever added between two
vertices for summations.  However, in some problems, the relations do
exist and can be important.

In this commit, the canonicalization is improved in a compromised way,
where the relations are taken into account in an approximate way to
minimize the performance penalty on problems without relation among the
summations.