Improve canonical order of summations with relations
For summations with the same colour, now we require that the unbounded
flavour comes before the bounded ones.  Among the bounded ones, those
with involvement of other summation dummies comes before those with.
The multi-set of free variables in the bounds are put as the last entry
of the colour.  This gives more aesthetically pleasing ordering the
summations after the canonicalization.