Improve simplification of deltas between free symbols
Previously, when a Kronecker delta does not involve any summed dummy, it
will simply be skipped for further simplification.  Now when this
happens and the operands of the delta are both of a simple linear form,
some simplification is also attempted to be performed to normalize the
usage of the two involved free symbols in the amplitude.