Add simplification utility for Wigner 3j nonzero rule
The three m quantum numbers in a Wigner 3j symbol need to sum to zero in
order for the symbol to be nonvanishing.  Now this utility can be used
for the simplification of other factors multiplying a product of Wigner
3j symbols.

Due to its relative complexity, this internal utility is also unit
tested in this commit.