Improve m phase in m contraction checking
In the previous code angular momentum simplification, the normal and
inverted contraction are only checked for the relative phase, with the
actual phase of the two m symbols left ambiguous.  With this commit,
this ambiguity is removed, the checker also tries to put the m symbols
into a preferred phase.