Generalize simplification based on Wigner 3j sum rules
This facility is now renamed power of negative one (pono)
simplification.  In addition to the simplification based on sum rules
based in the Wigner 3j style, the integral coefficients are also
canonicalized based on the integer/half-integer properties of the
coefficients.

However, the exact way these two kinds of simplifications play with each
other still needs to be investigated in order to reach a true canonical
form.