Implement simultaneous batch substituions for Tensor.subst_all
When ``simult_all`` is set to be true, subst_all performs
simultaneous batch substituions for a tensor instance, meaning
that the order of the substituion list does not change the
resulting tensor.

Tests of spin flipping and in-place BCS transformation are added
to the test set. More tests should be added later.