Add Dockerfile for building an image for drudge
The docker image is based on the base image for the drudge stack.  Only
drudge source code needs to be copied into it, compiled, and installed.
A swap directory is added for acting as the working directory.  This
directory is expected to be mounted a directory on the host machine for
the input and output files.